requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
tensorflow==2.20.0
pillow==11.3.0
numpy==2.2.6
requests==2.32.5
python-multipart==0.0.20
